css/css-tables/tables-ng/colgroup_col.html

<!doctype html>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src="/resources/check-layout-th.js"></script>
<script src="./support/table-inspector.js"></script>
<title>COLGROUP/COL</title>
<link rel="stylesheet" type="text/css" href="./support/tables-ng.css">
<link rel="author" title="Aleks Totic" href="atotic@chromium.org" />
<link rel="help" href="https://html.spec.whatwg.org/multipage/tables.html#the-colgroup-element" />
<link rel="help" href="https://html.spec.whatwg.org/multipage/tables.html#the-col-element" />
<link rel="help" href="https://stackoverflow.com/questions/29411447/what-does-the-width-attribute-of-colgroup-actually-specify" />
<style>
  table {
    background: gray;
  }

  td {
    background: #BFB;
  }

  td > div {
    display: inline-block;
    background: rgba(56,162,56,0.3);
  }
</style>
<h1>Colgroup col widths</h1>
<ol>
  <li>colgroup's purpose is to "group" cols</li>
  <li>one can think of colgroup as always generating max(colgroup.span,1) cols.</li>
  <li>colgroup css_width specifies width of generated cols.</li>
  <li>col.css_width can override colgroup width</li>
</ol>
<p class="testdesc">colgroups with spans</p>
<table>
  <colgroup width="100px"></colgroup>
  <colgroup width="100px" span=3></colgroup>
  <td data-expected-width=100>cg1</td>
  <td data-expected-width=100>cg2</td>
  <td data-expected-width=100>cg2</td>
  <td data-expected-width=100>cg2</td>
  <td data-expected-width=50><div style="width:50px">50px</div></td>
</table>

<p class="testdesc">cols with spans</p>
<table>
  <col width="100px"></colgroup>
  <col width="100px" span=3></colgroup>
  <td data-expected-width=100>col1</td>
  <td data-expected-width=100>col2</td>
  <td data-expected-width=100>col2</td>
  <td data-expected-width=100>col2</td>
  <td data-expected-width=50><div style="width:50px">50px</div></td>
</table>

<p class="testdesc">colgroups with cols</p>
<table>
  <colgroup span=4 style="width:100px">
    <col>
    <col style="width:50px">
    <col style="width:150px">
  </colgroup>
  <td data-expected-width=100>cg</td>
  <td data-expected-width=50>col1</td>
  <td data-expected-width=150>col2</td>
  <td data-expected-width=50><div style="width:50px">50px</div></td>
</table>

<p class="testdesc">colgroups with cols with spans</p>
<p class="error">Legacy fails col with span inside colgroup test.</p>
<table>
  <colgroup span=3 style="width:100px">
    <col>
    <col style="width:50px" span=2>
  </colgroup>
  <colgroup style="width:66px">
    <col span=2>
  </colgroup>
  <td data-expected-width=100>cg1</td>
  <td data-expected-width=50>col1</td>
  <td data-expected-width=50>col1</td>
  <td data-expected-width=66>cg2</td>
  <td data-expected-width=66>cg2</td>
  <td data-expected-width=75><div style="width:75px">50px</div></td>
</table>


<script>
  checkLayout("table");
</script>
